// src/interfaces.tsx
export interface record {
  name: string,
  address: string,
  age:number,
}

export interface user {
  id: number;
  name: string;
  username: string;
  address: address;
  phone: string;
  website: string;
  company: company;
}

interface address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo:geo
}

interface geo {
    lat: string;
    lng: string
}
interface company {
    name: string;
    catchPhrase: string;
    bs: string
}